// source/plugins/github/types/ActivityData.ts
export interface ActivityData {
  totalCommitContributions: number
  totalRepositoriesWithContributedCommits: number
  totalPullRequestContributions: number
  totalPullRequestReviewContributions: number
  totalIssueContributions: number
  restrictedContributionsCount: number
  repositoriesContributedTo: number
  issueComments: number
  organizations: number
  following: number
  sponsorshipsAsSponsor: number
  starredRepositories: number
  watching: number
  pullRequests: number
  issues: number
  gists: number
  projects: number
}